skills/smart-calendar/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-calendar"
version = "0.1.0"
description = "基于 Markdown 的个人日程管理工具"
requires-python = ">=3.10"
dependencies = [
    "python-frontmatter>=1.1.0",
    "dateparser>=1.2.0",
    "pendulum>=3.0.0",
    "july>=0.1.3",
    "matplotlib>=3.8.0",
    "rich>=13.0.0",
    "chinese-calendar>=1.9.0",
    "playwright>=1.40.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
sc = "smart_calendar.cli:main"

[tool.setuptools.packages.find]
include = ["smart_calendar*"]

[tool.setuptools.package-data]
"smart_calendar.render" = ["templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
